修复自动更新功能。

+++ app/src/main/java/org/cloud/panzer/mvp/contract/SplashContract.java
@@ -18,7 +18,7 @@ public interface SplashContract {
         void showDownloadSuccess(File file);
         void showDownloadFail();
         void showDownloadStart();
-        void showDownloadProgress(int progress, long total);
+        void showDownloadProgress(int progress, int total);
     }
 
     interface Model extends IModel {
